==> Code/Step 1.5 Create table of summary values for visualization.ipynb <==
import pandas as pd
import inspect
from fopt_toolkit import fopt_toolkit as fp
for i in inspect.getmembers(fp, inspect.isfunction):
    print i[0]

metadata_updated = pd.read_csv('../Metadata/project_metadata_updated.csv', dtype={'Date': object, 'Rep':object}, skiprows=0, delimiter= '\t')

metadata_updated.columns

#'Analysis_Type' --> a (absorption) or c (attenuation)
# Sample_Type --> fil, cal, raw
# Temp --> lalala
# ID
# Site

list_of_dfs = []
for index, row in metadata_updated.iterrows():
    ID, File, Site, Sample_Type, Analysis_Type, Date, Rep, Temp = row['ID'], row['Summary_file_path'], row['Site'], row['Sample_Type'],row['Analysis_Type'], row['Date'], row['Rep'], row['Temp']
    df = pd.read_csv(File, skiprows = 0, delimiter= '\t')         
    df.reset_index(inplace=True, drop=False)
    df.rename(columns = {'c_mean':'mean', 'a_mean':'mean', 'c_std':'std','a_std':'std'}, inplace = True)
    df = df[['wl','mean']]
    df['Temp'] = Temp
    df['ID'] = ID
    df['Site'] = Site
    df['Sample_Type'] = Sample_Type
    df['Analysis_Type'] = Analysis_Type
    df['Date'] = Date
    df['Rep'] = Rep
    list_of_dfs.append(df)

all_data = pd.concat(list_of_dfs)

new_filepath = fp.make_dir('Data/2_0_summary_stats_single_file/')

all_data.to_csv(new_filepath+'all_data.csv', sep=',',index=False)
